adds prelink.rs
